+++ R/interpolate.R
@@ -31,23 +31,19 @@
 #' # Or allow interpolate to find them in the current environment:
 #' topic <- "applies"
 #' interpolate(joke)
+#'
+#'
 interpolate <- function(prompt, ..., .envir = parent.frame()) {
   check_string(prompt)
 
-  if (!dots_named(...)) {
+  dots <- list2(...)
+  if (length(dots) > 0 && !is_named(dots)) {
     cli::cli_abort("All elements of `...` must be named")
   }
 
-  out <- glue::glue(prompt, ..., .open = "{{", .close = "}}", .envir = .envir)
-
-  if (length(out) != 1) {
-    cli::cli_abort(c(
-      "Must generate a single string.",
-      i = "Did you accidentally include a vector in {.arg ...}`?"
-    ))
-  }
-
-  out
+  envir <- list2env(dots, parent = .envir)
+  out <- glue::glue(prompt, .open = "{{", .close = "}}", .envir = envir)
+  ellmer_prompt(out)
 }
 
 #' @param path A path to a prompt file (often a `.md`).
@@ -74,6 +70,69 @@ interpolate_package <- function(
 read_file <- function(path) {
   file_contents <- readChar(path, file.size(path))
 }
+
+# Prompt class -----------------------------------------------------------------
+
+ellmer_prompt <- function(x) {
+  structure(x, class = c("ellmer_prompt", "character"))
+}
+
+is_prompt <- function(x) {
+  inherits(x, "ellmer_prompt")
+}
+
+#' @export
+print.ellmer_prompt <- function(
+  x,
+  ...,
+  max_items = 20,
+  max_lines = max_items * 10
+) {
+  n <- length(x)
+  n_extra <- length(x) - max_items
+  if (n_extra > 0) {
+    x <- x[seq_len(max_items)]
+  }
+
+  if (length(x) == 0) {
+    cli::cli_inform(c(x = "Zero-length prompt.\n"))
+    return(invisible(x))
+  }
+
+  bar <- if (cli::is_utf8_output()) "\u2502" else "|"
+
+  id <- format(paste0("[", seq_along(x), "] "), justify = "right")
+  indent <- paste0(cli::col_grey(id, bar), " ")
+  exdent <- paste0(strrep(" ", nchar(id[[1]])), cli::col_grey(bar), " ")
+
+  x[is.na(x)] <- cli::col_red("NA")
+  x <- paste0(indent, x)
+  x <- gsub("\n", paste0("\n", exdent), x)
+
+  lines <- strsplit(x, "\n")
+  ids <- rep(seq_along(x), length(lines))
+  lines <- unlist(lines)
+
+  if (length(lines) > max_lines) {
+    lines <- lines[seq_len(max_lines)]
+    lines <- c(lines, paste0(exdent, "..."))
+    n_extra <- n - ids[max_lines - 1]
+  }
+
+  cat(lines, sep = "\n")
+  if (n_extra > 0) {
+    cat("... and ", n_extra, " more.\n", sep = "")
+  }
+
+  invisible(x)
+}
+
+#' @export
+`[.ellmer_prompt` <- function(x, i, ...) {
+  ellmer_prompt(NextMethod())
+}
+
+# Helpers ----------------------------------------------------------------------
 
 # for mocking
 system.file <- NULL